Relocate and continue working on REST controller
