+++ src/test/java/com/colinmoerbe/javatodoapp/todo/TodoControllerTest.java
@@ -1,0 +1,31 @@
+package com.colinmoerbe.javatodoapp.todo;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class TodoControllerTest {
+
+    @BeforeEach
+    void setUp() {
+    }
+
+    @Test
+    void getTodos() {
+    }
+
+    @Test
+    void getTodo() {
+    }
+
+    @Test
+    void updateTodo() {
+    }
+
+    @Test
+    void createTodo() {
+    }
+
+    @Test
+    void deleteTodo() {
+    }
+}
